feat: add `replaceCallExpressions` util
