+++ codemods/array.prototype.copywithin/index.js
@@ -1,5 +1,8 @@
 import { ts } from '@ast-grep/napi';
-import { findDefaultImportIdentifier } from '../shared-ast-grep.js';
+import {
+	findDefaultImportIdentifier,
+	replaceCallExpressions,
+} from '../shared-ast-grep.js';
 
 const MODULE_NAME = 'array.prototype.copywithin';
 
@@ -30,28 +33,14 @@ export default function (options) {
 				return file.source;
 			}
 
-			const callExpressions = root.findAll({
-				rule: {
-					pattern: `${identifierName}($$$ARGS)`,
-				},
-			});
+			const callEdits = replaceCallExpressions(
+				root,
+				identifierName,
+				'copyWithin',
+				(args) => args.length >= 2,
+			);
 
-			for (const call of callExpressions) {
-				const argsMatch = call.getMultipleMatches('ARGS');
-				if (!argsMatch) continue;
-
-				const args = argsMatch.filter((m) => m.kind() !== ',');
-
-				if (args.length < 2) continue;
-
-				const arrayText = args[0].text();
-				const restArgs = args
-					.slice(1)
-					.map((m) => m.text())
-					.join(', ');
-
-				edits.push(call.replace(`${arrayText}.copyWithin(${restArgs})`));
-			}
+			edits.push(...callEdits);
 
 			for (const imp of imports) {
 				edits.push(imp.replace(''));
